:bug: Turning off listening to the pasteboard does not turn it on again
